Refactor info/version types into entropy-shared
