+++ crates/shared/src/lib.rs
@@ -22,6 +22,8 @@ pub use constants::*;
 pub use types::*;
 pub mod attestation;
 pub mod constants;
+#[cfg(feature = "production")]
+pub mod tss_node_info;
 pub mod types;
 
 #[cfg(any(feature = "user-native", feature = "user-wasm"))]
